Move the "query" property to the base class for Joining queries
